move common func IsP2 from account , core module to common module

+++ common/utils.go
@@ -16,6 +16,6 @@ func MasksHaveOverlap(m1, m2 uint32) bool {
 	return (m1 & bitMask) == (m2 & bitMask)
 }
 
-func IsP2(v uint64) bool {
+func IsP2(v uint32) bool {
 	return (v & (v - 1)) == 0
 }
